requirements.txt
# Backend
asgiref~=3.5.0
Django~=4.1.4
sqlparse~=0.4.2
psycopg2-binary==2.9.3
django-versatileimagefield~=2.2

# Files
whitenoise>=5.2,<5.3